[webcodecs] Clone VideoFrame when serializing
This changes the serialization semantics of VideoFrame so that the
lifetimes of the serialized (source) frame and deserialized (received)
frames are separate. The source and received frames must both be closed.

In the case that there are multiple receivers (eg. BroadcastChannel),
they still share a single handle, and only one of them should be closed.
This is necessary to ensure that no dangling references remain that
would require waiting for GC to destroy the underlying frames.

Bug: 1178038
Change-Id: I2199246b9d7caf9fd317b3c1dc8ae310ced9662d